html/browsers/browsing-the-web/navigating-across-documents/005.html

<!doctype html>
<title>Link with onclick navigation is aboprted by href navigation </title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<iframe id="test" name="test"></iframe>
<a target="test" onclick="document.getElementById('test').contentWindow.location='click.html'" href="href.html">Test</a>
<script>
/*---
specid: navigating-across-documents
steps: [6]
info: |
  Cancel any preexisting but not yet mature attempt to navigate browsingContext, including canceling any instances of the fetch algorithm started by those attempts. If one of those attempts has already created and initialized a new Document object, abort that Document also. (Navigation attempts that have matured already have session history entries, and are therefore handled during the update the session history with the new page algorithm, later.)
description: |
  1. Creates a srs-less iframe with an `id` of `test` and `name` of `test`
  2. Creates an anchor element with three attrs:
    a. `target` attr set to `test`, making the iframe this link's target
    b. `onclick` handler that selects the iframe in the DOM and sets its `location` to `click.html`.
    c. `href` attr with a value of `href.html`
  3. href.thml contains script that postmessages back the string "href"
  4. click.html contains script that postmessages back the string "click"
  5. 005.html then contains an async test which
    a. clicks the Link
    b. listens for the post message response
  6. `onclick` handlers get evaluated before hrefs, since the href is second, it is brought to maturity, and onclick's location change should be aborted per step 6 of 7.8.1
---*/
var t = async_test();
t.step(function() {
  document.getElementsByTagName("a")[0].click();
});
onmessage = t.step_func(
  function(e) {
    assert_equals(e.data, "href");
    t.done();
  });
</script>
